ICU-1722 rbnf can't handle small numbers
X-SVN-Rev: 14738

+++ icu4c/source/i18n/nfrule.h
@@ -91,7 +91,7 @@ private:
 
 private:
     int64_t baseValue;
-    int16_t radix;
+    int32_t radix;
     int16_t exponent;
     UnicodeString ruleText;
     NFSubstitution* sub1;
